tests: Add additional tests for export
This should round out the test coverage for export.
