+++ redskyctl/internal/commands/export/export_test.go
@@ -1,114 +1,44 @@
+/*
+Copyright 2020 GramLabs, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package export
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"net/http"
 	"os"
-	"os/exec"
 	"testing"
 
 	redsky "github.com/redskyops/redskyops-controller/api/v1beta1"
 	"github.com/redskyops/redskyops-controller/internal/config"
 	"github.com/redskyops/redskyops-controller/internal/server"
-	expapi "github.com/redskyops/redskyops-controller/redskyapi/experiments/v1alpha1"
+	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/experiments"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	"sigs.k8s.io/yaml"
 )
-
-/*
-func TestExport(t *testing.T) {
-	samplePatch := `spec:
-  template:
-    spec:
-      containers:
-        - name: postgres
-          resources:
-          limits:
-            cpu: "{{ .Values.cpu }}m"
-            memory: "{{ .Values.memory }}Mi"
-          requests:
-            cpu: "{{ .Values.cpu }}m"
-            memory: "{{ .Values.memory }}Mi"`
-
-	sampleExperiment := &redsky.Experiment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "sampleExperiment",
-			Namespace: "default",
-		},
-		Spec: redsky.ExperimentSpec{
-			Parameters: []redsky.Parameter{},
-			Metrics:    []redsky.Metric{},
-			Patches: []redsky.PatchTemplate{
-				{
-					TargetRef: &corev1.ObjectReference{
-						Kind:       "Deployment",
-						APIVersion: "apps/v1",
-						Name:       "postgres",
-					},
-					Patch: samplePatch,
-				},
-			},
-			TrialTemplate: redsky.TrialTemplateSpec{},
-		},
-		Status: redsky.ExperimentStatus{},
-	}
-
-	tmpfile, err := ioutil.TempFile("", "experiment")
-	require.NoError(t, err)
-	defer os.Remove(tmpfile.Name())
-
-	b, err := json.Marshal(sampleExperiment)
-	require.NoError(t, err)
-	_, err = tmpfile.Write(b)
-	require.NoError(t, err)
-
-	// TODO add filename: parameter so we can mock
-	cfg := &config.RedSkyConfig{}
-	cfg.Load()
-
-	testCases := []struct {
-		desc string
-		args []string
-	}{
-		{
-			desc: "default",
-			args: []string{
-				"als-368",
-				"--filename",
-				tmpfile.Name(),
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(fmt.Sprintf("%q", tc.desc), func(t *testing.T) {
-
-			var b bytes.Buffer
-			cmd := NewCommand(cfg)
-			cmd.SetOut(&b)
-
-			if len(tc.args) > 0 {
-				cmd.SetArgs(tc.args)
-			}
-
-			err := cmd.Execute()
-			assert.NoError(t, err)
-			log.Println(b.String())
-		})
-	}
-}
-*/
 
 func TestReadExperiment(t *testing.T) {
 	_, tmpfile := createTempExperimentFile(t)
 	defer os.Remove(tmpfile.Name())
 
-	// TODO add filename: parameter so we can mock
 	cfg := &config.RedSkyConfig{}
 	cfg.Load()
 
@@ -139,6 +69,52 @@ func TestReadExperiment(t *testing.T) {
 			}
 
 			assert.NoError(t, err)
+		})
+	}
+}
+
+func TestLookupTrial(t *testing.T) {
+	exp, tmpfile := createTempExperimentFile(t)
+	defer os.Remove(tmpfile.Name())
+
+	testCases := []struct {
+		desc          string
+		expectedError bool
+		e             *ExportCommand
+		trialName     string
+	}{
+		{
+			desc:          "default",
+			expectedError: false,
+			e:             &ExportCommand{RedSkyAPI: &fakeRedSkyServer{}, experiment: exp},
+			trialName:     "test-1234",
+		},
+		{
+			desc:          "no api",
+			expectedError: true,
+			e:             &ExportCommand{experiment: exp},
+		},
+		{
+			desc:          "bad trial",
+			expectedError: true,
+			e:             &ExportCommand{RedSkyAPI: &fakeRedSkyServer{}, experiment: exp},
+			trialName:     "test-test",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(fmt.Sprintf("%q", tc.desc), func(t *testing.T) {
+			trialNames, err := experiments.ParseNames(append([]string{"trials", tc.trialName}))
+			assert.NoError(t, err)
+
+			trial, err := tc.e.GetTrial(context.Background(), trialNames[0])
+			if tc.expectedError {
+				assert.Error(t, err)
+				return
+			}
+
+			assert.NoError(t, err)
+			assert.Equal(t, len(trial.Spec.Assignments), 2)
 		})
 	}
 }
@@ -177,92 +153,71 @@ func TestPatches(t *testing.T) {
 			pnt, err := tc.e.Patches(context.Background())
 			assert.NoError(t, err)
 			assert.Equal(t, len(pnt), 1)
+
 			// TODO this name doesnt seem right, need to get that sorted
 			assert.Equal(t, "-Deployment-postgres", pnt[0].TargetName)
 		})
 	}
 }
 
-func createTempExperimentFile(t *testing.T) (*redsky.Experiment, *os.File) {
-	samplePatch := `spec:
-  template:
-    spec:
-      containers:
-        - name: postgres
-          resources:
-          limits:
-            cpu: "{{ .Values.cpu }}m"
-            memory: "{{ .Values.memory }}Mi"
-          requests:
-            cpu: "{{ .Values.cpu }}m"
-            memory: "{{ .Values.memory }}Mi"`
+func TestRenderPatch(t *testing.T) {
+	exp, tmpfile := createTempExperimentFile(t)
+	defer os.Remove(tmpfile.Name())
 
-	sampleExperiment := &redsky.Experiment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "sampleExperiment",
-			Namespace: "default",
-		},
-		Spec: redsky.ExperimentSpec{
-			Parameters: []redsky.Parameter{},
-			Metrics:    []redsky.Metric{},
-			Patches: []redsky.PatchTemplate{
-				{
-					TargetRef: &corev1.ObjectReference{
-						Kind:       "Deployment",
-						APIVersion: "apps/v1",
-						Name:       "postgres",
-					},
-					Patch: samplePatch,
-				},
-			},
-			TrialTemplate: redsky.TrialTemplateSpec{},
-		},
-		Status: redsky.ExperimentStatus{},
+	trial := &redsky.Trial{}
+	server.ToClusterTrial(trial, &wannabeTrial.TrialAssignments)
+	trial.ObjectMeta.Labels = map[string]string{
+		redsky.LabelExperiment: "test",
 	}
 
-	tmpfile, err := ioutil.TempFile("", "experiment")
+	pgfile, err := ioutil.TempFile("", "deployment")
 	require.NoError(t, err)
 
-	b, err := json.Marshal(sampleExperiment)
+	_, err = pgfile.Write(pgDeployment)
 	require.NoError(t, err)
+	defer os.Remove(pgfile.Name())
 
-	_, err = tmpfile.Write(b)
-	require.NoError(t, err)
-
-	return sampleExperiment, tmpfile
-}
-
-var wannabeTrial = expapi.TrialItem{
-	// This is our target trial
-	TrialAssignments: expapi.TrialAssignments{
-		Assignments: []expapi.Assignment{
-			{
-				ParameterName: "cpu",
-				Value:         "100",
-			},
-			{
-				ParameterName: "memory",
-				Value:         "200",
-			},
+	testCases := []struct {
+		desc          string
+		expectedError bool
+		e             *ExportCommand
+	}{
+		{
+			desc:          "default",
+			expectedError: false,
+			e:             &ExportCommand{Config: &fakeClient{filename: pgfile.Name()}, trial: trial, experiment: exp},
 		},
-	},
-	Number: 1234,
-	Status: expapi.TrialCompleted,
-}
+	}
 
-type fakeClient struct {
-	filename string
-}
+	for _, tc := range testCases {
+		t.Run(fmt.Sprintf("%q", tc.desc), func(t *testing.T) {
+			pnt, err := tc.e.Patches(context.Background())
+			assert.NoError(t, err)
+			assert.Equal(t, len(pnt), 1)
 
-func (f *fakeClient) Endpoints() (config.Endpoints, error) { return nil, nil }
+			output, err := patchResource(pnt)
+			assert.NoError(t, err)
 
-func (f *fakeClient) Authorize(ctx context.Context, transport http.RoundTripper) (http.RoundTripper, error) {
-	return nil, nil
-}
+			// With []byte from output, we should try to turn it into an actual deployment
+			deployment := &appsv1.Deployment{}
+			err = yaml.Unmarshal(output, deployment)
+			assert.NoError(t, err)
 
-func (f *fakeClient) SystemNamespace() (string, error) { return "", nil }
+			limits := deployment.Spec.Template.Spec.Containers[0].Resources.Limits
+			/*
+				yolo := limits[corev1.ResourceCPU]
+				fmt.Println((&yolo).MilliValue())
+				oloy := limits[corev1.ResourceMemory]
+				fmt.Println((&oloy).MilliValue())
+			*/
 
-func (f *fakeClient) Kubectl(ctx context.Context, arg ...string) (*exec.Cmd, error) {
-	cmd := exec.CommandContext(ctx, "/usr/bin/paste", f.filename)
-	return cmd, nil
+			// wannabe doesnt contain the actual units ( m/Mi/Gi/etc )
+			// but since we're testing and know it, we'll just hardcode
+			expectedCPU := resource.MustParse(fmt.Sprintf("%s%s", wannabeTrial.TrialAssignments.Assignments[0].Value, "m"))
+			expectedMemory := resource.MustParse(fmt.Sprintf("%s%s", wannabeTrial.TrialAssignments.Assignments[1].Value, "Mi"))
+
+			assert.True(t, limits[corev1.ResourceCPU].Equal(expectedCPU))
+			assert.True(t, limits[corev1.ResourceMemory].Equal(expectedMemory))
+		})
+	}
 }
