+++ redskyctl/internal/commands/export/patcher.go
@@ -1,3 +1,19 @@
+/*
+Copyright 2020 GramLabs, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package export
 
 import (
@@ -7,27 +23,13 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-func createPatch(patchOp *redsky.PatchOperation) (patchData []byte, err error) {
-	switch patchOp.PatchType {
-	case types.JSONPatchType:
-		// Can use kustomize here
-	case types.MergePatchType:
-	// TODO ?
-	case types.StrategicMergePatchType:
-		// Can use kustomize here
-	case types.ApplyPatchType:
-		// TODO ?
-	}
-
-	// TODO: for offline testing
-	output := pgDeployment
+func createPatch(patchOp *redsky.PatchOperation, targetObj []byte) (patchData []byte, err error) {
 	pOM := &metav1.PartialObjectMetadata{}
-	if err = json.Unmarshal(output, pOM); err != nil {
+	if err = json.Unmarshal(targetObj, pOM); err != nil {
 		return patchData, err
 	}
 
@@ -36,7 +38,7 @@ func createPatch(patchOp *redsky.PatchOperation) (patchData []byte, err error) {
 	// object to render the patch on top of base gvk object.
 	// This is necessary so we can create a properly structured patch object
 	// to apply via kustomize(?)
-	obj, err := runtime.Decode(unstructured.UnstructuredJSONScheme, output)
+	obj, err := runtime.Decode(unstructured.UnstructuredJSONScheme, targetObj)
 	if err != nil {
 		return patchData, err
 	}
@@ -50,6 +52,7 @@ func createPatch(patchOp *redsky.PatchOperation) (patchData []byte, err error) {
 	if err = json.Unmarshal(patchOp.Data, &spec); err != nil {
 		return patchData, err
 	}
+
 	u := &unstructured.Unstructured{}
 	u.SetUnstructuredContent(spec)
 	u.SetName(patchOp.TargetRef.Name)
@@ -62,12 +65,4 @@ func createPatch(patchOp *redsky.PatchOperation) (patchData []byte, err error) {
 	}
 
 	return patchData, err
-}
-
-func createStrategicMergePatch(po *redsky.PatchOperation) {
-
-}
-
-func createMergePatch(po *redsky.PatchOperation) {
-
 }
